compression: Move integer encoder implementation
Move deltafor and deltafor_column to the compression module.
