+++ src/v/cloud_storage/remote_segment_index.h
@@ -12,12 +12,12 @@
 
 #include "bytes/iobuf.h"
 #include "bytes/iobuf_parser.h"
+#include "compression/delta_for.h"
 #include "model/fundamental.h"
 #include "model/record_batch_types.h"
 #include "seastarx.h"
 #include "storage/parser.h"
 #include "units.h"
-#include "utils/delta_for.h"
 
 #include <seastar/util/log.hh>
