+++ src/v/cloud_storage/segment_meta_cstore.cc
@@ -11,12 +11,12 @@
 #include "cloud_storage/segment_meta_cstore.h"
 
 #include "cloud_storage/types.h"
+#include "compression/delta_for.h"
 #include "config/configuration.h"
 #include "model/fundamental.h"
 #include "model/metadata.h"
 #include "model/timestamp.h"
 #include "serde/serde.h"
-#include "utils/delta_for.h"
 
 #include <bitset>
 #include <exception>
