+++ src/v/cloud_storage/tests/segment_meta_cstore_bench.cc
@@ -9,10 +9,10 @@
  */
 
 #include "cloud_storage/segment_meta_cstore.h"
+#include "compression/delta_for.h"
 #include "model/fundamental.h"
 #include "random/generators.h"
 #include "seastarx.h"
-#include "utils/delta_for.h"
 
 #include <seastar/testing/perf_tests.hh>
 #include <seastar/util/defer.hh>
