+++ src/v/compression/delta_for.h
