+++ src/v/compression/tests/delta_for_characterization_data.cc
